+++ docs/CHANGELOG.rst
@@ -16,6 +16,7 @@ Added
 * Amplicon descriptor schemas
 * Amplicon report generator
 * Add Rattus norvegicus organism choice to sample schema
+* Transforming form Phred 64 to Phred 33 when uploading fastq reads
 
 Fixed
 -----